Config.in: add option to ban textrels
Ideally, this option would default to yes when a musl-based toolchain is
used, as musl doesn't support TEXTRELs[1] and programs with TEXTRELs
will crash on start-up under musl.

However, to avoid a large build fallout, it defaults to no for now.
After this option has been enabled in build tests for a while and some
of the fallout has been fixed, the default can be changed.

[1]: https://www.openwall.com/lists/musl/2020/09/25/4