system: Add option to maintain a supervision tree
One of the major drawbacks of classic init/rc systems is that dead
daemons remain dead: Once started by the rc scripts, they are left to
look after themselves.

A simple approach to keep them running reliably is to supervise them:
In a tree with PID 1 at its root, every process knows which children
ought to run under it, and it respawns them accordingly. This is the
supervision tree.

Classic init already provides rudimentary supervision in the form of
inittab. This patch uses inittab to establish a more flexible
supervision tree based on runit's rvsvdir or s6's s6-svscan.

Signed-off-by: J. Neuschäfer <j.neuschaefer@gmx.net>